refactor: split the structure watchdog and cycle types out of tokenCycle
